Removed the code that tried to calculate branch
Originally the script tried to determine the git branch programatically
using git.  There is a problem with this approach when using a CI system
like Jenkins because the repo is in a detatched head state.  This means
the branch determination is always returning HEAD.  The code now takes
the requested environment type and computes the branch name.

--development === "devlop"
--testing === "master"
--production == package.json semver